This commit makes adding and deleting from a PrimarySource async, and pops
2008-04-09  Gabriel Burt  <gabriel.burt@gmail.com>

	This commit makes adding and deleting from a PrimarySource async, and pops
	up UserJobs if those operations take very long.  Very importantly, it also
	adds the ability to copy the track model selection so slow running jobs
	can process it w/o risking it changing on them.

	* src/Core/Banshee.Services/Banshee.Collection.Database/DatabaseTrackListModel.cs:
	Handle having no artist or album model gracefully.

	* src/Core/Banshee.Services/Banshee.Playlist/PlaylistSource.cs: Fixed up
	Add/Delete methods.

	* src/Core/Banshee.Services/Banshee.ServiceStack/UserJob.cs: Handle
	multiple Finish calls gracefully.

	* src/Core/Banshee.Services/Banshee.Sources/DatabaseSource.cs: New
	AddAllTracks method for use by MergeSource method.

	* src/Core/Banshee.Services/Banshee.Sources/PrimarySource.cs: Change slow
	operations to use a new CachedList object instead of iterating over the
	model's Selection, since it or the model might change underneath us.  Add
	UserJobs for adding/deleting.

	* src/Core/Banshee.Services/Banshee.Sources/Source.cs: Don't raise
	UserNotifyUpdated if we're the active source.

	* src/Core/Banshee.Services/Makefile.am: New files

	* src/Core/Banshee.Services/Banshee.Collection.Database/CachedList.cs: New
	class that contains a SqliteModelCache, so it's backed by CoreCache and
	can be used to store info (eg the selected tracks from a source) so they
	can be operated on independent of the original selection/model.

	* src/Core/Banshee.Services/Banshee.ServiceStack/BatchUserJob.cs: UserJob
	convenience subclass.

	* src/Core/Banshee.ThickClient/Banshee.Sources.Gui/SourceView_DragAndDrop.cs:
	Do MergeSource operations in a thread.  Should probably be done in
	PrimarySource.

	* src/Dap/Banshee.Dap.MassStorage/Banshee.Dap.MassStorage/MassStorageSource.cs:
	Whitespace.

	* src/Extensions/Banshee.PlayQueue/Banshee.PlayQueue/PlayQueueSource.cs:
	Get rid of OnUserNotifyUpdated, this is done in DatabaseSource now.

	* src/Libraries/Hyena/Hyena.Collections/RangeCollection.cs: Add Count
	property to Range to avoid doing End - Start + 1 everywhere.

	* src/Libraries/Hyena/Hyena.Data.Sqlite/SqliteModelCache.cs: Make work
	without a Selection.

svn path=/trunk/banshee/; revision=3743

+++ src/Libraries/Hyena/Hyena.Data.Sqlite/SqliteModelCache.cs
@@ -94,7 +94,10 @@ namespace Hyena.Data.Sqlite
             }
 
             uid = FindOrCreateCacheModelId (String.Format ("{0}-{1}", uuid, typeof(T).Name));
-            selection_uid = FindOrCreateCacheModelId (String.Format ("{0}-{1}-Selection", uuid, typeof(T).Name));
+
+            if (model.Selection != null) {
+                selection_uid = FindOrCreateCacheModelId (String.Format ("{0}-{1}-Selection", uuid, typeof(T).Name));
+            }
 
             if (model.CachesJoinTableEntries) {
                 select_range_command = new HyenaSqliteCommand (
@@ -172,19 +175,21 @@ namespace Hyena.Data.Sqlite
                 )
             );
 
-            delete_selection_command = new HyenaSqliteCommand (String.Format (
-                "DELETE FROM {0} WHERE ModelID = {1}", CacheTableName, selection_uid
-            ));
-            
-            save_selection_command = new HyenaSqliteCommand (String.Format (
-                "INSERT INTO {0} (ModelID, ItemID) SELECT {1}, ItemID FROM {0} WHERE ModelID = {2} LIMIT ?, ?",
-                CacheTableName, selection_uid, uid
-            ));
+            if (model.Selection != null) {
+                delete_selection_command = new HyenaSqliteCommand (String.Format (
+                    "DELETE FROM {0} WHERE ModelID = {1}", CacheTableName, selection_uid
+                ));
+                
+                save_selection_command = new HyenaSqliteCommand (String.Format (
+                    "INSERT INTO {0} (ModelID, ItemID) SELECT {1}, ItemID FROM {0} WHERE ModelID = {2} LIMIT ?, ?",
+                    CacheTableName, selection_uid, uid
+                ));
 
-            get_selection_command = new HyenaSqliteCommand (String.Format (
-                "SELECT OrderID FROM {0} WHERE ModelID = {1} AND ItemID IN (SELECT ItemID FROM {0} WHERE ModelID = {2})",
-                CacheTableName, uid, selection_uid
-            ));
+                get_selection_command = new HyenaSqliteCommand (String.Format (
+                    "SELECT OrderID FROM {0} WHERE ModelID = {1} AND ItemID IN (SELECT ItemID FROM {0} WHERE ModelID = {2})",
+                    CacheTableName, uid, selection_uid
+                ));
+            }
         }
 
         private bool has_select_all_item = false;
@@ -245,9 +250,9 @@ namespace Hyena.Data.Sqlite
         public override void Reload ()
         {
             lock (this) {
-                if (last_reload_fragment != model.ReloadFragment) {
+                if (last_reload_fragment != model.ReloadFragment || last_reload_command == null) {
                     last_reload_fragment = model.ReloadFragment;
-                    last_reload_command = new HyenaSqliteCommand (reload_sql + last_reload_fragment);
+                    last_reload_command = new HyenaSqliteCommand (String.Format ("{0}{1}", reload_sql, last_reload_fragment));
                 }
 
                 Clear ();
@@ -267,7 +272,9 @@ namespace Hyena.Data.Sqlite
         private bool saved_selection = false;
         public void SaveSelection ()
         {
-            if (model.Selection.Count > 0 && !(has_select_all_item && model.Selection.AllSelected)) {
+            if (model.Selection != null && model.Selection.Count > 0 &&
+                !(has_select_all_item && model.Selection.AllSelected))
+            {
                 connection.Execute (delete_selection_command);
                 saved_selection = true;
